Q3G7HM8 dissolve scrubber/hooks: usePersistedFlag is generic and leaves the scrubber, the appearance hooks join layout
