+++ source/gui/client/components/TimeScrubber.tsx
@@ -43,11 +43,11 @@ import {
 	segmentAt,
 	useExitTransition,
 	isPeriodWindow,
-	usePersistedFlag,
 	useNarrowBar,
 	usePrefersReducedMotion,
 	windowNamesIssues,
 } from '../lib/scrubber';
+import {usePersistedFlag} from '../lib/use-persisted-flag';
 import {canPlayTimeline} from '../lib/theatre';
 import {HintContent, ScrubberLayout} from './ScrubberLayout';
 import {ScatterLayer, ScatterPoint} from './ScatterCanvas';
